+++ src/COV_Allier_RIL.cpp
@@ -34,7 +34,7 @@ SEXP cpp_calculate_covariance_RIL_allier(
     const List& genMap,
     const NumericMatrix& M,
     const NumericMatrix& U,
-    double t,
+    int t,
     double intensity,
     const NumericVector& weights,
     bool covariance = false,
